Revert "Add periodic sync"
