--- projectBlueWater/src/main/java/com/lasthopesoftware/bluewater/client/stored/service/adapter/DummyAuthenticator.java
@@ -1,48 +1,0 @@
-package com.lasthopesoftware.bluewater.client.stored.service.adapter;
-
-import android.accounts.AbstractAccountAuthenticator;
-import android.accounts.Account;
-import android.accounts.AccountAuthenticatorResponse;
-import android.content.Context;
-import android.os.Bundle;
-
-public class DummyAuthenticator extends AbstractAccountAuthenticator {
-	public DummyAuthenticator(Context context) {
-		super(context);
-	}
-
-	@Override
-	public Bundle editProperties(AccountAuthenticatorResponse response, String accountType) {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public Bundle addAccount(AccountAuthenticatorResponse response, String accountType, String authTokenType, String[] requiredFeatures, Bundle options) {
-		return null;
-	}
-
-	@Override
-	public Bundle confirmCredentials(AccountAuthenticatorResponse response, Account account, Bundle options) {
-		return null;
-	}
-
-	@Override
-	public Bundle getAuthToken(AccountAuthenticatorResponse response, Account account, String authTokenType, Bundle options) {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public String getAuthTokenLabel(String authTokenType) {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public Bundle updateCredentials(AccountAuthenticatorResponse response, Account account, String authTokenType, Bundle options) {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public Bundle hasFeatures(AccountAuthenticatorResponse response, Account account, String[] features) {
-		throw new UnsupportedOperationException();
-	}
-}
