--- projectBlueWater/src/main/java/com/lasthopesoftware/bluewater/client/stored/service/adapter/EmptyContentProvider.java
@@ -1,43 +1,0 @@
-package com.lasthopesoftware.bluewater.client.stored.service.adapter;
-
-import android.content.ContentProvider;
-import android.content.ContentValues;
-import android.database.Cursor;
-import android.net.Uri;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-
-public class EmptyContentProvider extends ContentProvider {
-	@Override
-	public boolean onCreate() {
-		return true;
-	}
-
-	@Nullable
-	@Override
-	public Cursor query(@NonNull Uri uri, @Nullable String[] projection, @Nullable String selection, @Nullable String[] selectionArgs, @Nullable String sortOrder) {
-		return null;
-	}
-
-	@Nullable
-	@Override
-	public String getType(@NonNull Uri uri) {
-		return null;
-	}
-
-	@Nullable
-	@Override
-	public Uri insert(@NonNull Uri uri, @Nullable ContentValues values) {
-		return null;
-	}
-
-	@Override
-	public int delete(@NonNull Uri uri, @Nullable String selection, @Nullable String[] selectionArgs) {
-		return 0;
-	}
-
-	@Override
-	public int update(@NonNull Uri uri, @Nullable ContentValues values, @Nullable String selection, @Nullable String[] selectionArgs) {
-		return 0;
-	}
-}
