--- projectBlueWater/src/main/res/xml/authenticator.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<account-authenticator xmlns:android="http://schemas.android.com/apk/res/android"
-                   android:accountType="namehillsoftware.com"
-                   android:icon="@drawable/clearstream_launcher"
-                   android:smallIcon="@drawable/clearstream_launcher"
-                   android:label="@string/app_name" />
