--- projectBlueWater/src/main/res/xml/syncadapter.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<sync-adapter
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        android:contentAuthority="com.lasthopesoftware.bluewater.provider"
-        android:accountType="namehillsoftware.com"
-        android:userVisible="false"
-        android:supportsUploading="false"
-        android:allowParallelSyncs="false"
-        android:isAlwaysSyncable="true"/>
